Require latest task definition to be running by default
- describe target task's task definition
- describe the deployments; require exactly 1 active deployment
- fail if the current deployment is not running the latest task
  defintion

Fast rollbacks are two phases: rollback the task definition to the
previous review, then revert the change from github. The time between
the task definition rollback and the revert deploy is lengthy and a
deploy of the service during that time before the revert PR is merged
will reintroduce the bug that was rolled back. So, by default, require
that the service is running the latest task definition before the deploy
is executed.